ZOOKEEPER-965 - Add initial database work to support multi ops.
This is for ZOOKEEPER-965: Need a multi-update command to allow
multiple znodes to be updated safely. This commit includes all the
changes necessary to PrepRequestProcessor, SyncRequestProcessor,
FinalRequestProcessor, DataTree, as well as many, many other classes.

I've now gotten all the unit tests that Ted originally wrote to pass.
I've also added a bunch for unit tests into MultiTransactionTest.

This work is *not* complete IMO, but it is pretty close (I think).
There are a few things I wanted to work out better but it wasn't
working out as well as I'd have liked. Specifically:

1. No client work is done. I'm not sure where to start on this, so if
someone wants to point me the right direction I'll tackle that next.

2. PrepRequestProcessor.java:432: I don't like how I had to serialize
the individual Txns into the MultiTxn only to deserialize it in the
next request processor in DataTree. I wanted to just have MultiTxn
contain a list of Txns. However, the jute auto code generation stuff
got in my way. I couldn't figure out how to make the MultiTxn contain a
list of Txns so I made it a buffer that I serialize into/out of. It
works, but I think this could be better.

3. Several places where we do a switch on types of ops inside multi op,
and in default it throws an IOException. That
doesn't seem very appropriate. Maybe instead a KeeperException.APIERROR
or perhaps we need a (new) specific exception?

4. There is currently no way to tell which op in a multi op failed.
It's implemented so that it throws the KeeperException corresponding to
the first op in the multi op that caused the multi op to pass. However,
it might be more useful if there was a way to know exactly which op
failed. I tried several approaches to implement this but couldn't get
it working correctly. Maybe someone can offer some advice on this.