ZOOKEEPER-965 - Enhanced multi-op so caller can tell which op failed.
This turned out to be a lot harder than it looked. The jist of
the problem was I was throwing an exception early in
PrepRequestProcessor on the first error in the multi op. That was
the right thing to do, but it didn't allow returning the results
back to the caller as the exception caused us to discard the current
multi-op and create a new ErrorTxn that traversed all the other
request handlers.

Instead, this new implementation retains the multi-op all the way
through all the request handlers (though it becomes a MultiTxn
and ultimately a MultiResponse). It now allows returning back
to the caller the result for each individual Op.

In ZooKeeper's multi function, it still ultimately behaves the
same as before -- namely the first error in the results array
gets detected as the fatal error that caused the multi op to fail
and this is what we throw as our exception.

The major difference is that we first (optionally) populate a given
result array *before* we throw the exception.

There are some rough details here still and some things I'd like to
fix. But it works really well and all the major bugs are flushed
out so I'm committing....