ZOOKEEPER-965 - Fix bug in error reporting on failed multi-op.
Multi-op can now store off results into passed in array. If the
multi-op failed then every entry in that result array will be of
type ErrorResult (which implements OpResult).

For all Ops that *would* have succeeded (e.g. those before the failing
op in the multi op) the error code for those inside the ErrorResult
is 0. For the *ONE* failing op, the error code will be the real
error code that caused the failure. This can be then translated into
a proper KeeperException (e.g. NONODE) to figure out why the op
failed. For all ops after the failing op in the multi op, the error
in the ErrorResult will be RUNTIMEINCONSISTENCY. Since the point of
the multi-op is to preserve multi-op atomic consistency, I chose this
error code to denote that the ops after the failing op weren't attempted
at all to avoid a runtime inconsistency.

We may change these values later, but this is working great for now.

Anyhow, I digress greatly... The bug being fixed here was that it was
reporting the first ErrorResult as the fatal/failing op in the multiop
and throwing that error. But it has to look for one that has a non-zero
error code (based on above implementation details).