ZOOKEEPER-965 - Refactor sync_completion out of zookeeper_process
Moved the gigantic "else" statement from zookeeper_process that
processed sync_completions. This is because I will need it to
process completions inside multi-op list of completions so it
needs to be a standalone function that can be called from both
places.

Only real change was to move the logic that checked return code
inside each case statement to instead be checked in the caller.
Greatly simplified code with no functional change.