ZOOKEEPER-965 - Refactored op init mechanism.
Previously I was using a struct with a union of all the op types. This
worked well on C as you can specify which field of union to initialize
to using extended struct initializers. This isn't supported on C++
without special compiler flags.

So I opted for a much simpler solution. Put all the necessary arguments
into the main op_t structure and provide some simple macros that provide
simple convenience wrappers to only take the arguments appropriate for
that op (and in the same order as the non-multi version). This provides
very simple caller syntax with no added overhead or memory allocations.