ZOOKEEPER-965 - Fix incomplete CheckVersion in Multi Op.
The java server-side for the CheckVersionRequest was never fully flushed
out. While implementing the C API side of things I discovered all the
places this needed to be fixed on the server side.

The C API allows passing in an optional pointer for the Stat structure
to get back the stat of the node in question. Or just pass in NULL if
you only care about doing the version check itself.