ZOOKEEPER-965 - Fix CHECK_OP to work in a non-multi context.
The CHECK_OP wasn't implemented to work in a non-multi context. It was added
specifically for multi-op, but for completeness and consistency I felt it was
better to go ahead and support it in a non-multi context.

Also added some unit test code around the multi-op version of a CHECK_OP as
well as a non-multi op versio nof CHECK_OP.

Also fixed/added comments to zookeper.h regarding new APIs.