ZOOKEEPER-965 - Fix error code handling.
Error codes were not propagating out from the server properly.
The original design dictated that if the multi op fails, then
all the entries in the multi result would be an error result. The
error code in that error result would be used to determine which
op failed in the multi-op. Specifically, error codes in the multi result:

0: The op would have succeeded had the entire multi op not failed
ZRUNTIMEINCONSISTENCY: The op was AFTER the failing op and was aborted
Anything else: The real error code for the op that caused the entire
multi op to fail (this is the one that is thrown as well).

Also, discovered another error code issue. If a op in the multi
op is after the failing op, it still ran through PrepRequestProcessor.
That was confusing (and unecessary) as we know the entire multi op
is going to fail. Modified it to abort all ops after a failing op
to avoid unecessary work and also to avoid confusion in the log files.

Also added some unit tests to verify all above described changes.