ZOOKEEPER-965 - Refactor zoo_op_t to use a union.
Refactoring code in zookeeper.h/zookeeper.c and associated unit
test to use an anonymous union for all the various types of ops
that can be in a multi-op. This tightens the C interface
considerable without adversely affecting the useability.

Added convenience init routines to initialize the right fields
in the zoo_op_t based on op type.