ZOOKEEPER-965 - fix inoperable multi called from watch context.
While integrating the new zoo_multi and zoo_amulti into the
code at our company, we discovered a nasty bug in the C API. If
a multi op is triggered from a watch callback, then it would hang
indefinitely. This was due to a fundamental flaw in my earlier
implementation whereby I treated all the ops in a multi op as
asynchronous, with a forced synchronous tail completion that I
tacked onto the end of my completion list. BUT, that tail completion
ultimately never got called when called from a watch context!!

The fix was ultimately to create code in both process functions
(both sync and async) for dealing with multi-ops. Each op inside
the multiop is dealt with the same, so they call into a common
function. This way, all the existing logic for dealing with
sync/async completions now deals with multi ops instead of
me trying to deal with that specially.

I've also added a new unit test that demonstrated this bug and
also now passes with my refactoring.